pyproject.toml
[project]
name = "langchain_and_qdrant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.26",
    "langchain-groq>=0.3.4",
    "langchain-ollama>=0.3.3",
    "langchain-qdrant>=0.2.0",
    "pypdf>=5.7.0",
    "streamlit>=1.46.1",
    "torch>=2.7.1",
    "ultralytics>=8.3.161",
]
